Finished EAHELPER05
When you need to print out a mode, register stored in the bits 0-5 (this
would typically be this first parameter in like a move d2,d3 it would be
the d2) simply call EAHELPER05. It will print out whatever belongs in
that first parameter.

NOTE: Immediate mode implemented, but untested